backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coverage==7.6.1
decorator==4.4.2
exceptiongroup==1.3.1
fastapi==0.124.4
h11==0.16.0
idna==3.11
imageio==2.35.1
imageio-ffmpeg==0.5.1
iniconfig==2.1.0
moviepy==1.0.3
numpy==1.24.4
opencv-python==4.12.0.88
packaging==25.0
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.12
pydantic==2.10.6
pydantic-core==2.27.2
pytest==8.3.5
pytest-cov==5.0.0
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.4
sniffio==1.3.1
starlette==0.44.0
tomli==2.3.0
tqdm==4.67.1
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
gunicorn==21.2.0
